+++ server/api.php
@@ -2,7 +2,7 @@
 
 $DEBUG = true;
 
-$VERSION = "0.1.0+0419261208";
+$VERSION = "0.1.0+0419261241";
 
 require_once("dbconfig.php");